src/xadmin/requirements.txt
six~=1.16.0
future~=0.18.2
django-crispy-forms~=1.14.0
django-formtools
httplib2~=0.20.4
django-import-export